+++ packages/design-system-web/src/components/geov-carousel/geov-carousel.stories.tsx
@@ -1,8 +1,9 @@
+import React from 'react';
 import { Components } from '../..';
 import { GeovCarousel } from '../../../.storybook/stencil-generated/component';
 
 export default {
-  title: 'Components/Carousel',
+  title: 'Design Components/Carousel',
   component: GeovCarousel,
 };
 const Template = (args: Components.GeovCarousel) => <GeovCarousel style={{height:400, maxWidth:600}} {...args}></GeovCarousel>;