+++ packages/design-system-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.3-pr-21.2 (2022-09-28)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.3.3-pr-21.1 (2022-09-28)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.3.3-pr-21.0 (2022-09-28)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
 ## 1.3.2 (2022-09-19)
 
 **Note:** Version bump only for package @geovistory/design-system-react